dom/base/test/test_meta_viewport_initial_scale_2.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>initial-scale=2 in meta viewport</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="initial-scale=2">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>initial-scale=2</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function initial_scale_2() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      is(info.width, 400, "width should be scaled by 1 / initial-scale");
      is(info.height, 240, "height should be scaled by 1 / initial-scale");
      is(info.defaultZoom, 2, "initial-scale should be 2");
    });
  </script>
</body>
</html>
